Completely redesigned the ObjectFactory in MBSim. - Registring new object can (should) now be done locally where the class is defined instead of centrally in objectfactory.cc. - No initialzation of object factories from modules needed (This enables easy plugin concepts for modules (issue 30)) - No need to define a createABC function for each type (ABC) of object to be create (automatically done using templates) - Creating of arbitaty container types using automatic, type safe, casts using templates
Moreover this will make the ObjectFactory much easier to use and multiple fmatvec::Vector types for a single XML Element name
will be possible later (currently even not possible but will follow soon) with this approach. This will remove the Vec3, Vec<Ref>
and Vec<Var> differentiation in the Function[123]_* concept (at least in XML).

Please note the comment in the new files dummyof.cc. These files will be removed later if all type registrations have been moved
to their local places in the cc file where the corresonding class is defined.
This is currently not done to be able to fastly change all registered types during test the new object factory in a single file
instead of changes many files.