First checkin for new structure
The new structure should be more be like in MBSim with contact
kinematics and and a force law. This is not working right now properly
as a "mesh" still exists which fullfills force law tasks as well as
kinematical ones.
Furthermore the example is not running which could depend on different
stuff --> TODO