Fixed and improved ConstantFunction.
ConstantFunction<Ret(...)> uses now as ctor ConstantFunction<Ret(...)>(Ret x)
instead of ConstantFunction<Ret(...)>(double x)!
This is more general and obvioas. However, this requires some changes in
the examples (since e.g. ConstantFunction<VecV(...)>(3) is interpretet by
fmatvec as ConstantFunction<VecV(...)>(VecV(3)) and hence as a vector of
dimention 3 instead of a vector of dimension 1 with the value of 3).