Provide a flag to switch to a constraint solver within the plot-routine.
Using an event-driven time integration scheme the constraints are determined by a
solver for linear equations within the plot-routine. For indeterminate sytems this might result in incorrect contat forces, as the contact laws are not considered. Hence, we provide a flag in order to switch to a constraint solver within the plot-routine.