Revised plot feature.
Introduced a PlotFeatureEnum class. An instance of this class
on e.g. MBSim namespace level introduces an plot feature enum.
Use this new "enums" everywhere. This way each c++ plotfeature
enum is encaplusated by the module c++ namespace and cannot be
in conflict with any other plot feature enum.

For XML use a EnumFactory is introduced which maps a XML QName
(represented by FQN) to the c++ enum value (instance of the
PlotFeatureEnum class). All c++ enums must be registered by the
EnumFactory or this (just like already done for the ObjectFactory).