Removed the static system variable from Integrator.
This static variable was preventing all integrators from being
instantiated more than once per process. This prevents e.g. to use more
than one MBSim cosim FMU in a importing tool.

A Solver stores now the system by Solver::setSystem(...) and the "this"
pointer of the solver object is passed as user data to the Fortran integrators
which pass is back to the e.g. static RHS callback functions.

RKSuiteIntegrator is the only which does not support this. Hence, still
a static variable is use here and a check for double instantation of
RKSuiteIntegrator was added.