Enable reduced form of system of nonlinear equations
If the differential system has the special structure that dq(i)/dt = u(i), the system of nolinear equations to be solved by implicit integrators can be reduced. For the implicit Euler integrator this saves significant computing time. Also the RADAU5 integrator provides this feature, however an improvement can not be observed there.