Removed all singleton MBSim modelling classes (Environment)
The MBSim::Environment was implemented as a singleton (only one instance
per type e.g. MBSimEnvironment, HydraulicEnvironment). This is relaxed
now to only one instance per type per DynamicSystemSolver.
This allows e.g. to hold more than on DynamicSystemSolver object in
memory which use different Environments. This already may happen for FMI
exports if more than one MBSim FMU is loaded into a system. Before this
change the Environment settings in FMU 1 may have overwritten the
settings in FMU 2 and the other way around. This is not expected and a
now go.
This is fixed now but required several changes even some changes in each
source example.

However, this also removed the need for singletons at all which
simplfies the ObjectFactory.