Added a internal state vector
DynamicSystemSolver holds now a vector for internal states.
Its size is defined by the sum of internal state vector sizes of all
elements (for now just Link's and DynamicSystem's).
Each element has a vector reference to this internal state vector (like
all other states e.g. u, v, la, ...)
This internal state is availalbe as two variants:
- the current internal state (curis): this is the internal state being used
- the next state (nextis): if a element updates its internal state it
  stores the new internal state in nextis.
The integrator is responsible to copy nextis to curis after each
successfull integration step, but NOT after rejected steps.
This way the internal state cannot be polluted by calls from rejected
steps which may represent significantly wrong states.

This is used now e.g. to store the starting value for the local root finding
in contact kinematics class CircleSpatialContour. If this is not done it
may happen that a integrator step, which is rejected later on, changes the
starting value such that another local minima is found. Now the reject
of the integrator step also reverts the internal state, being the
starting value.