Added FilteredFrameObserver
It creates an OpenMBV::Frame which moves analog to a referenced
MBSim::Frame.
However, single translations/rotations can be set to a constant value.
This is very useful in OpenMBV to move the camera with this filtered
frame if e.g. the camera should just translate with an MBSim::Frame but
not rotate.