Use automatic refresh of openmbv.
So we no longer need to call explicitly refresh in mbsimgui since openmbv now
automatically reloads when the h5 file is written again.

Fixed also a memory leak in mbsimflatxml when an exception is thrown.
This is importend to call the dtors of HDF5Serie which ensures a valid shared memory.