Fixed mbsim.xsd schema
The XML schema mbsim.xsd included a bug which makes mbsim.xsd invalid
according the XML schema standard!
It worked in mbsim since the currently used version of xerces does not
detect this bug in the schema file (and a workaround is contained in mbsim.xsd).
Fixing this bug is only possible by breaking the XML interface of mbsim!
This is done with this commit.
For now the old and new definition is supported (the old prints a deprecated message)