Added plotAttributes to Element
To each Element plot attribures can now be added.
These attributes are just static data in form of key/value pairs which
are attached to the dataset/group in the plot of the corresponding
Element.

They can be used e.g. to mark datasets which keywords for automatic
processing (searching) in post processing tools.