Removed the "emit widgetChanged()" from all initializeUsingXML code
The order of the XML elements may be different to the order of the
dependencies of corresponding GUI widgets. E.g.

<generalizedInitialPosition> comes before <stateDependentTranslation>
but the "generalized initial position" widget depends on the
"state dependent translation" (the vector length of inipos depends on
the defined translation element)
For this reason after initializeUsingXML no widgetChange signal can
be emitted.