Enable Embed for all "container" elements.
Since some time mbsimgui can handle UnknownObject's for all containers
to enable the GUI also for object of unknown XML (unknown to the
mbsimgui but not unknown to mbsimxml).
This also allows to use the Embed element in all these container.
Hence, Embed is now allowed in the XML schema for all "container" (all
<xs:element ref="..."/> elements).

One container in mbsimgui was missing and is also added with this
commit.

Especially all Function's can now use Embed.