Call a links stopvector only if svSize it > 0.
This is needed to enable elements which do not check for the svSize
before settings the sv-vector. This may only happen for element which
use a sv for set-valued force laws but do not use a sv for single-values
force laws but to check check this when e.g. updateStopVector is called.