Allow LinearTransferSystem with a x size of 0 and allow acceleration inputs in this case.
There is no real need to disallow a LinearTransferSystem with a x size of 0 (a pure feedthrough).
Hence, its allowed now.

In this case it also allowed that the input is a acceleration
signal if the output of LinearTransferSystem is not passed back to the RHS of the
simulation.
This allows to use LinearTransferSystem for a feedthrough calculation
for observers.