Improved performance and fixes for RADAU5
The Jacobian part with respect to lambda (la) and
with respect to the GGL algebraic variables (l) are done now analytically:

du/dla = evalM()^-1 * evalJrla(), dREST/dla = 0

dq/dl = evalW(), dREST/dl = 0

Hence, only the Jac wrt q, u and x must be done numerically which
improves the performance a lot if the system has many lambda's.

Fixed DAE1 mode if evalT()!=E

Fixed reduced from if x states exist

Set mass matrix to 1 or 0 respectively: setting to 0 was missing but the
mass matrix is not gurantied to be init with 0 by the integrator code
(at least not for none dialog matrices even so this was not added with this commit)

Disallow GGL with reduced form since GGL will always add a none diagonal
part (W*) to the
qd = T * u + W' * l
equation even when T=E

This fixes allow to use RADAU5 for move problems. Tests have been added.