Improved the responsibility of mbsimgui
A "update" and "simulation" call creates till now a single XML file with
inline Embed elements for all embeds.
Than MBXMLUtils preprocess is called on this XML code and
then mbsimflatxml is called as a seperate process.

The new approach not a single XML file with
inline Embed element for all embeds is created.
Than mbsimxml is called as a seperate process on this XML file also does
the preprocessing.
This way the preprocessing is done in a seperate process and the code
returns earlier to the Qt event loop which makes the GUI more
responsible to user interactions.