Improved system shift for RADAU5 integreator.
DAE integrators are usually very sensitive regarding proper initial
values: if the initial state is not consistent these integrators fail to start.

This means that the following conditions must hold at integration start:

ODE formalism:
no special requirements since the integrator does not "see" the
constraints at all (the DSS solves these on acceleration level)

DAE1 formalism:
a) the constraint forces lambda must be consistent (lambda's are part of the integrator state)

DAE2 formalism:
a) same as above
b) consistent velocities must be given

DAE3/GGL formalism:
a) same as above
b) same as above
c) consistent position must be given

a) is already ensured for RADAU5 since we explicitly calculate lambda
(calling evalla()) before the integration starts in radau5_integrator.cc

b) and c) can be ensured for all integrators by setting the DSS flag <initialProjection> to true.

However, a integrator start also happens after a root/event was found by
the integrator. But DSS::shift() does not always call a projection
of the positions and velocites, this depends on the type of the root (getRootID()).
Hence, shift() was now extended to be able to report if a projection was
done. Integrators can then react on this flag and make the required
projections before (re)starting the integrator.

This solves problems with hanging DAE integrators after roots which do
not require a projection but the drift is already too large for the
integrators starting phase.