Re-Revert "Do not use state and time dependent functions as parent"
This re-reverts commit 8b7bf90970ce2cd5838dd348b63fa5e4661ff6b2.
Hence, commit bf3138bdcffc34c03fdae74989635d52a4be0cb2 is active again.

This introduces a backward-incompatiblity if a function inside a c++
TimeDependentFunction or StateDependentFunction references another
element using the XML attribute ref="..." notation.

Either the XML code of test models must be adapted accordingly (removing
of one '../' at the beginning of a relative reference,

or a legacy handling of such elements must be implemented which accept
both variants: the old (invalid) reference with an additional '../' and
the new (valid) reference where the correct number of '../' is used.

Such a legacy handling is possible e.g. by catching the error if the
reference could not be found (by getByPath) and then adapting the reference
(remove of the '../' in front) and retry the reference resolution
(using getByPath)